% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swingeR.R
\name{print_breeding_groups}
\alias{print_breeding_groups}
\title{Print the results of swingeR() in a readable format}
\usage{
print_breeding_groups(results)
}
\arguments{
\item{results}{Output of \code{\link{swingeR}}.}
}
\value{
The combination of breeding groups with everage relatednes/kinship per group and combination from swingeR resutls.
}
\description{
Print the results of swingeR() in a readable format
}
\examples{
\dontrun{
BreedGroups <- print_breeding_groups(outputSwingeR)
}
}
