% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swingeR.R
\name{print_matrix_groups}
\alias{print_matrix_groups}
\title{Print relatedness/kinship matrix of the combination of groups selected by swingeR().}
\usage{
print_matrix_groups(outputswinger, kr_df)
}
\arguments{
\item{outputswinger}{Output of \code{\link{swingeR}}.}

\item{kr_df}{original relatedness/kinship data frame used as input for \code{\link{swingeR}}.}
}
\value{
A set of pairwise kinship/relatedness matrix for each group in the swingeR resutls.
}
\description{
Print relatedness/kinship matrix of the combination of groups selected by swingeR().
}
\examples{
\dontrun{
myBredGMatrix <- print_matrix_groups(outputswinger, kinship_df)
}
}
