% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swingeR.R
\name{print_breeding_groups}
\alias{print_breeding_groups}
\title{Print the results of swingeR() in a readable format}
\usage{
print_breeding_groups(results)
}
\arguments{
\item{results}{Output of \code{\link{swingeR}}.}
}
\value{
Invisibly \code{NULL}; called for its printed output.
}
\description{
Print the results of swingeR() in a readable format
}
